+++ src/Heartbeat/Heartbeat.php
@@ -6,6 +6,7 @@ namespace OrcaServices\Heartbeat\Heartbeat;
 use Cake\Chronos\Chronos;
 use Cake\Collection\Collection;
 use Cake\Core\Configure;
+use OrcaServices\Heartbeat\Heartbeat\Sensor\Severity;
 use OrcaServices\Heartbeat\Heartbeat\Sensor\Status;
 
 /**
@@ -32,12 +33,12 @@ class Heartbeat
      */
     public function check()
     {
-        $sensors = $this->_getEnabledSensors();
+        $sensors = $this->getEnabledSensors();
 
         $this->sensorStatuses = [];
         foreach ($sensors as $sensorName => $sensorConfig) {
-            $sensor = $this->_getSensor($sensorName, $sensorConfig);
-            $this->sensorStatuses[] = $sensor->getStatus();
+            $sensor = $this->getSensor($sensorName, $sensorConfig);
+            $this->sensorStatuses[] = $sensor->getSensorStatus();
         }
 
         return $this;
@@ -48,7 +49,7 @@ class Heartbeat
      *
      * @return array All enabled sensors.
      */
-    protected function _getEnabledSensors(): array
+    protected function getEnabledSensors(): array
     {
         $sensors = (array)Configure::read('App.Heartbeat.Sensors');
         $collection = collection($sensors);
@@ -66,7 +67,7 @@ class Heartbeat
      * @param array $sensorConfig The sensor configuration.
      * @return Sensor The configures sensor.
      */
-    protected function _getSensor(string $sensorName, array $sensorConfig): Sensor
+    protected function getSensor(string $sensorName, array $sensorConfig): Sensor
     {
         $config = new Sensor\Config($sensorName, $sensorConfig);
         if (!$this->cached) {
@@ -100,8 +101,8 @@ class Heartbeat
 
         $systemStatus = !$sensorStatuses->some(function ($sensorStatus) {
             /** @var Status $sensorStatus */
-            if ($sensorStatus->getSeverity() === Status::STATUS_CRITICAL) {
-                return $sensorStatus->getStatus() === false;
+            if ($sensorStatus->isCritical()) {
+                return $sensorStatus->status === false;
             }
 
             return false;
@@ -114,7 +115,7 @@ class Heartbeat
             $systemStatus,
             0, // TODO Calculate the duration for the whole heartbeat
             Chronos::now(),
-            Status::STATUS_CRITICAL
+            Severity::CRITICAL,
         );
     }