+++ src/Heartbeat/Sensor.php
@@ -43,18 +43,18 @@ abstract class Sensor
     }
 
     /**
-     * Get the status
+     * Get the sensor status
      *
      * @return Sensor\Status The sensor status.
      */
-    public function getStatus(): Status
+    public function getSensorStatus(): Status
     {
-        $cachedStatus = $this->_getCachedStatus();
+        $cachedStatus = $this->getCachedStatus();
         if ($cachedStatus !== false) {
             return $cachedStatus;
         }
 
-        return $this->_getNonCachedStatus();
+        return $this->getNonCachedStatus();
     }
 
     /**
@@ -62,13 +62,13 @@ abstract class Sensor
      *
      * Resets the cache, if disabled.
      *
-     * @return bool|Status The cached status or false.
+     * @return Status|bool The cached status or false.
      */
-    protected function _getCachedStatus()
+    protected function getCachedStatus(): Status|bool
     {
         $sensorCaching = $this->config->getCached();
 
-        $this->_resetCacheConfig($sensorCaching);
+        $this->resetCacheConfig($sensorCaching);
 
         $cacheKey = self::CACHE_NAME . '_' . strtolower(Text::slug($this->config->getName()));
         if ($sensorCaching === false) {
@@ -87,7 +87,7 @@ abstract class Sensor
             return $cachedStatus;
         }
 
-        $nonCachedStatus = $this->_getNonCachedStatus();
+        $nonCachedStatus = $this->getNonCachedStatus();
         Cache::write($cacheKey, $nonCachedStatus, self::CACHE_NAME);
 
         return $nonCachedStatus;
@@ -96,10 +96,10 @@ abstract class Sensor
     /**
      * Reset the cache configuration
      *
-     * @param bool|string $sensorCaching The sensor cache configuration, either a bool or a relative time string.
+     * @param string|bool $sensorCaching The sensor cache configuration, either a bool or a relative time string.
      * @return void
      */
-    protected function _resetCacheConfig($sensorCaching): void
+    protected function resetCacheConfig(string|bool $sensorCaching): void
     {
         Cache::drop(self::CACHE_NAME);
 
@@ -110,7 +110,7 @@ abstract class Sensor
 
         $settings = array_merge(
             (array)Cache::getConfig('default'),
-            ['duration' => $duration, 'className' => 'File']
+            ['duration' => $duration, 'className' => 'File'],
         );
 
         Cache::setConfig(self::CACHE_NAME, $settings);
@@ -121,41 +121,39 @@ abstract class Sensor
      *
      * @return Status The status object.
      */
-    protected function _getNonCachedStatus(): Status
+    protected function getNonCachedStatus(): Status
     {
         $start = microtime(true);
-        $status = $this->_getStatus();
+        $status = $this->getStatus();
         $end = microtime(true);
 
         $duration = $end - $start;
         $duration = round($duration, 3);
 
-        $status = new Status(
+        return new Status(
             $this->config->getName(),
             $status,
             $duration,
             Chronos::now(),
-            $this->config->getSeverity()
+            $this->config->getSeverity(),
         );
-
-        return $status;
     }
 
     /**
      * Get the status
      *
-     * @return mixed The sensor status.
+     * @return bool The sensor status.
      */
-    abstract protected function _getStatus();
+    abstract protected function getStatus(): bool;
 
     /**
      * Get the value of the given setting or an optional fallback default value
      *
      * @param string $name The name of the setting to retrieve.
-     * @param null|mixed $default The optional default value, if the setting is not set.
+     * @param mixed|null $default The optional default value, if the setting is not set.
      * @return string|null The value of the setting or the provided default, if not set.
      */
-    protected function getSetting(string $name, $default = null): ?string
+    protected function getSetting(string $name, mixed $default = null): ?string
     {
         $settings = $this->config->getSettings();