+++ src/Heartbeat/Sensor/Config.php
@@ -13,20 +13,13 @@ class Config
 {
     use InstanceConfigTrait;
 
-    /** @var array All valid severity levels */
-    public const SEVERITY_LEVELS = [
-        Status::STATUS_CRITICAL,
-        Status::STATUS_NONCRITICAL,
-        Status::STATUS_INFORMATIONAL,
-    ];
-
     /**
      * The default config
      */
     protected array $_defaultConfig = [
         'name' => null,
         'enabled' => true,
-        'severity' => Status::STATUS_NONCRITICAL,
+        'severity' => Severity::NONCRITICAL,
         'class' => null,
         'cached' => false,
         'settings' => [],
@@ -93,24 +86,22 @@ class Config
     /**
      * Set the severity level
      *
-     * @param int $severity The severity level.
+     * @param Severity $severity The severity level.
      * @return void
-     * @throws InvalidArgumentException If an invalid severity level was given.
      */
-    public function setSeverity(int $severity): void
+    public function setSeverity(Severity $severity): void
     {
-        $this->assertSeverity($severity);
         $this->setConfig('severity', $severity, false);
     }
 
     /**
      * Get the severity level
      *
-     * @return int The severity level.
+     * @return Severity The severity level.
      */
-    public function getSeverity(): int
+    public function getSeverity(): Severity
     {
-        return (int)$this->getConfig('severity');
+        return $this->getConfig('severity');
     }
 
     /**
@@ -138,9 +129,9 @@ class Config
     /**
      * Get whether or how long the status should be cached
      *
-     * @return bool|string
+     * @return string|bool
      */
-    public function getCached()
+    public function getCached(): string|bool
     {
         return $this->getConfig('cached');
     }
@@ -148,14 +139,11 @@ class Config
     /**
      * Set whether or how long the status should be cached
      *
-     * @param bool|string $cached Whether or how long the status should be cached.
+     * @param string|bool $cached Whether or how long the status should be cached.
      * @return void
-     * @throws InvalidArgumentException If not a valid boolean or string was given.
-     * @todo Cover the exception.
      */
-    public function setCached($cached): void
+    public function setCached(string|bool $cached): void
     {
-        $this->assertCached($cached);
         $this->setConfig('cached', $cached, false);
     }
 
@@ -187,12 +175,12 @@ class Config
      * @return void
      * @throws InvalidArgumentException If an invalid severity level was given.
      */
-    private function assertSeverity($severity): void
+    private function assertSeverity(mixed $severity): void
     {
-        if (!in_array($severity, self::SEVERITY_LEVELS, true)) {
+        if (!$severity instanceof Severity) {
             throw new InvalidArgumentException(sprintf(
                 'Severity must be a valid severity level, got "%s" instead.',
-                $severity
+                is_object($severity) ? $severity::class : (string)$severity,
             ));
         }
     }
@@ -204,12 +192,12 @@ class Config
      * @return void
      * @throws InvalidArgumentException If an invalid boolean or string was given.
      */
-    private function assertCached($cached): void
+    private function assertCached(mixed $cached): void
     {
         if (!is_bool($cached) && !is_string($cached)) {
             throw new InvalidArgumentException(sprintf(
                 'Cached must be either a bool or a string, "%s" given instead',
-                $cached
+                $cached,
             ));
         }
     }