+++ src/HeartbeatPlugin.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+
+namespace OrcaServices\Heartbeat;
+
+use Cake\Core\BasePlugin;
+
+/**
+ * CakePHP Heartbeat Plugin
+ */
+class HeartbeatPlugin extends BasePlugin
+{
+}