+++ tests/TestCase/Heartbeat/Sensor/ConfigTest.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 namespace OrcaServices\Heartbeat\Test\TestCase\Heartbeat\Sensor;
 
 use Cake\TestSuite\TestCase;
+use InvalidArgumentException;
 use OrcaServices\Heartbeat\Heartbeat\Sensor\Config;
+use OrcaServices\Heartbeat\Heartbeat\Sensor\Severity;
 use OrcaServices\Heartbeat\Test\TestCase\Sensor\DummySensor;
 
 /**
@@ -35,11 +37,11 @@ class ConfigTest extends TestCase
         $sensorName = 'Dummy Sensor';
         $sensorConfigArray = [
             'enabled' => true,
-            'severity' => 1,
+            'severity' => Severity::INFORMATIONAL,
             'class' => DummySensor::class,
             'cached' => true,
             'settings' => [
-                'connection_name' => 'test',
+                'connection' => 'test',
             ],
         ];
         $sensorConfig = new Config($sensorName, $sensorConfigArray);
@@ -76,7 +78,7 @@ class ConfigTest extends TestCase
         $sensorConfig = new Config($sensorName, $sensorConfigArray);
         static::assertEquals($sensorName, $sensorConfig->getName());
         static::assertTrue($sensorConfig->getEnabled());
-        static::assertEquals(2, $sensorConfig->getSeverity());
+        static::assertEquals(Severity::NONCRITICAL, $sensorConfig->getSeverity());
         static::assertEquals($sensorConfigArray['class'], $sensorConfig->getClass());
         static::assertEquals([], $sensorConfig->getSettings());
         static::assertFalse($sensorConfig->getCached());
@@ -90,7 +92,7 @@ class ConfigTest extends TestCase
      */
     public function testInvalidSeverity(): void
     {
-        $this->expectException(\InvalidArgumentException::class);
+        $this->expectException(InvalidArgumentException::class);
         $sensorName = 'Dummy Sensor';
         $sensorConfigArray = [
             'enabled' => true,
@@ -108,11 +110,11 @@ class ConfigTest extends TestCase
      */
     public function testInvalidCached(): void
     {
-        $this->expectException(\InvalidArgumentException::class);
+        $this->expectException(InvalidArgumentException::class);
         $sensorName = 'Dummy Sensor';
         $sensorConfigArray = [
             'enabled' => true,
-            'severity' => 1,
+            'severity' => Severity::INFORMATIONAL,
             'class' => DummySensor::class,
             'cached' => 1,
         ];