+++ tests/TestCase/Heartbeat/SensorTest.php
@@ -5,9 +5,10 @@ namespace OrcaServices\Heartbeat\Test\TestCase\Heartbeat;
 
 use Cake\Chronos\Chronos;
 use Cake\TestSuite\TestCase;
-use OrcaServices\Heartbeat\Heartbeat\Sensor;
 use OrcaServices\Heartbeat\Heartbeat\Sensor\Config;
+use OrcaServices\Heartbeat\Heartbeat\Sensor\Severity;
 use OrcaServices\Heartbeat\Test\TestCase\Sensor\DummySensor;
+use ReflectionClass;
 
 /**
  * Sensor Test
@@ -21,14 +22,14 @@ class SensorTest extends TestCase
      *
      * @return void
      * @covers ::__construct
-     * @throws \ReflectionException
+     * @throws ReflectionException
      */
     public function testConstructor()
     {
         $sensorName = 'Dummy Sensor';
         $sensorConfig = [
             'enabled' => true,
-            'severity' => 1,
+            'severity' => Severity::INFORMATIONAL,
             'class' => DummySensor::class,
         ];
         $sensorConfig = new Config($sensorName, $sensorConfig);
@@ -42,8 +43,8 @@ class SensorTest extends TestCase
      * Tests the constructor
      *
      * @return void
+     * @covers ::getSensorStatus
      * @covers ::getStatus
-     * @covers ::_getStatus
      */
     public function testGetStatus()
     {
@@ -51,29 +52,28 @@ class SensorTest extends TestCase
         $sensorName = 'Dummy Sensor';
         $sensorConfig = [
             'enabled' => true,
-            'severity' => 1,
+            'severity' => Severity::INFORMATIONAL,
             'class' => DummySensor::class,
         ];
         $sensorConfig = new Config($sensorName, $sensorConfig);
         $sensorClassName = $sensorConfig->getClass();
         /** @var Sensor $sensor */
         $sensor = new $sensorClassName($sensorConfig);
-        $status = $sensor->getStatus();
-        $this->assertInstanceOf(Sensor\Status::class, $status);
-        $this->assertEquals('Dummy Sensor', $status->getName());
-        $this->assertTrue($status->getStatus());
-        $this->assertEquals(0, $status->getDuration());
-        $this->assertEquals('2017-03-30 12:45:37', $status->getLastExecuted());
-        $this->assertEquals(1, $status->getSeverity());
+        $sensorStatus = $sensor->getSensorStatus();
+        $this->assertEquals('Dummy Sensor', $sensorStatus->name);
+        $this->assertTrue($sensorStatus->status);
+        $this->assertEquals(0, $sensorStatus->duration);
+        $this->assertEquals('2017-03-30 12:45:37', $sensorStatus->lastExecuted);
+        $this->assertEquals(Severity::INFORMATIONAL, $sensorStatus->severity);
     }
 
     /**
      * Tests whether the check result was fetched from cache or by running the check now
      *
      * @return void
-     * @covers ::_getCachedStatus
-     * @covers ::_resetCacheConfig
-     * @covers ::_getNonCachedStatus
+     * @covers ::getCachedStatus
+     * @covers ::resetCacheConfig
+     * @covers ::getNonCachedStatus
      */
     public function testWasCheckCached()
     {
@@ -81,7 +81,7 @@ class SensorTest extends TestCase
         $sensorName = 'Cached Sensor';
         $sensorConfig = [
             'enabled' => true,
-            'severity' => 1,
+            'severity' => Severity::INFORMATIONAL,
             'class' => DummySensor::class,
             'cached' => '+1 seconds',
         ];
@@ -90,19 +90,19 @@ class SensorTest extends TestCase
 
         /** @var DummySensor $sensor */
         $sensor = new $sensorClassName($sensorConfig);
-        $status = $sensor->getStatus();
+        $sensorStatus = $sensor->getSensorStatus();
 
         // Assert that result was not cached
-        $this->assertFalse($status->wasCheckCached());
+        $this->assertFalse($sensorStatus->wasCheckCached());
 
         // Get status again and assert that result was cached
-        $status = $sensor->getStatus();
-        $this->assertTrue($status->wasCheckCached());
+        $sensorStatus = $sensor->getSensorStatus();
+        $this->assertTrue($sensorStatus->wasCheckCached());
 
         // Get status again after slightly more than a second and assert that result was not cached
         sleep(2);
-        $status = $sensor->getStatus();
-        $this->assertFalse($status->wasCheckCached());
+        $sensorStatus = $sensor->getSensorStatus();
+        $this->assertFalse($sensorStatus->wasCheckCached());
 
         //// Wait another second to let the cache be reset
         sleep(1);
@@ -112,9 +112,9 @@ class SensorTest extends TestCase
      * Tests whether the check result was fetched from cache when cache is disabled
      *
      * @return void
-     * @covers ::_getCachedStatus
-     * @covers ::_resetCacheConfig
-     * @covers ::_getNonCachedStatus
+     * @covers ::getCachedStatus
+     * @covers ::resetCacheConfig
+     * @covers ::getNonCachedStatus
      */
     public function testWasCheckCachedWhenCacheDisabled()
     {
@@ -122,7 +122,7 @@ class SensorTest extends TestCase
         $sensorName = 'Uncached Sensor';
         $sensorConfig = [
             'enabled' => true,
-            'severity' => 1,
+            'severity' => Severity::INFORMATIONAL,
             'class' => DummySensor::class,
             'cached' => false,
         ];
@@ -131,11 +131,11 @@ class SensorTest extends TestCase
 
         /** @var DummySensor $sensor */
         $sensor = new $sensorClassName($sensorConfig);
-        $status = $sensor->getStatus();
+        $sensorStatus = $sensor->getSensorStatus();
 
-        $this->assertFalse($status->wasCheckCached());
-        $status = $sensor->getStatus();
-        $this->assertFalse($status->wasCheckCached());
+        $this->assertFalse($sensorStatus->wasCheckCached());
+        $sensorStatus = $sensor->getSensorStatus();
+        $this->assertFalse($sensorStatus->wasCheckCached());
     }
 
     /**
@@ -144,13 +144,12 @@ class SensorTest extends TestCase
      * @param mixed $object The object
      * @param string $property The property name
      * @return mixed The value
-     * @throws \ReflectionException
+     * @throws ReflectionException
      */
-    public static function getProperty($object, string $property)
+    public static function getProperty(mixed $object, string $property): mixed
     {
-        $reflectedClass = new \ReflectionClass($object);
+        $reflectedClass = new ReflectionClass($object);
         $reflection = $reflectedClass->getProperty($property);
-        $reflection->setAccessible(true);
 
         return $reflection->getValue($object);
     }